removed support.txt, cleared the hardcoded variables, modularization done
