+++ .gitignore
@@ -1,1 +1,2 @@
 escalation-ninja-tests.go
+support.txt
