+++ escalation-ninja.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
@@ -14,293 +12,6 @@ import (
 )
 
 // :x: Failed to create channel: Slack API error: invalid_name_specials - there might be the chance it's importing colons or it might be breaking because the jira is in upper case
-func fixChannelName(name string) string {
-	name = strings.ToLower(name)              // Convert to lowercase
-	name = strings.ReplaceAll(name, "_", "-") // Replace underscores with dashes
-	name = strings.ReplaceAll(name, " ", "-") // Replace spaces with dashes
-	name = strings.ReplaceAll(name, ".", "-") // Replace dots with dashes
-	name = strings.ReplaceAll(name, "@", "-") // Replace @ with dashes
-	name = strings.ReplaceAll(name, "#", "-") // Replace # with dashes
-	name = strings.ReplaceAll(name, ":", "-") // Replace colons with dashes
-	return name
-}
-
-// will need the SlackRequest to be handled as an object - I think it'll be easier
-type SlackRequest struct {
-	Token       string `json:"token"`
-	TeamID      string `json:"team_id"`
-	TeamDomain  string `json:"team_domain"`
-	ChannelID   string `json:"channel_id"`
-	ChannelName string `json:"channel_name"`
-	UserID      string `json:"user_id"`
-	UserName    string `json:"user_name"`
-	Command     string `json:"command"`
-	Text        string `json:"text"`
-	ResponseURL string `json:"response_url"`
-	TriggerID   string `json:"trigger_id"`
-}
-
-// TEMP: Struct for Slack API response
-type SlackChannelResponse struct {
-	OK      bool `json:"ok"`
-	Channel struct {
-		ID   string `json:"id"`
-		Name string `json:"name"`
-	} `json:"channel"`
-	Error string `json:"error,omitempty"`
-}
-
-// creating a function to handle the Slack Channel creation
-func createSlackChannel(channelName string) (string, error) {
-	slackToken := os.Getenv("SLACK_BOT_TOKEN")
-	fmt.Println("DEBUG - Slack Token:", slackToken) //checking the Token, as I got a "Token not set" error
-	if slackToken == "" {
-		return "", fmt.Errorf("Slack token not set")
-	}
-
-	slackApiUrl := "https://slack.com/api/conversations.create"
-
-	//Requesting payload
-	payload := map[string]string{
-		"name": channelName,
-	}
-
-	//payload to JSON
-	payloadBytes, _ := json.Marshal(payload)
-
-	//HTTP request
-	req, err := http.NewRequest("POST", slackApiUrl, bytes.NewBuffer(payloadBytes))
-	if err != nil {
-		return "", err
-	}
-
-	//setting headers
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+slackToken)
-
-	//executing the request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	//Decode response
-	var slackResp SlackChannelResponse
-	json.NewDecoder(resp.Body).Decode(&slackResp)
-
-	//printing the response for debugging since I got a "unknown_method" error
-	fmt.Println("DEBUG - Slack API Response:", slackResp)
-
-	//Handling the API response
-	if !slackResp.OK {
-		return "", fmt.Errorf("Slack API error: %s", slackResp.Error)
-	}
-
-	log.Printf("✅ Channel Create: #%s (ID: %s)", slackResp.Channel.Name, slackResp.Channel.ID)
-	return slackResp.Channel.ID, nil
-}
-
-// creating a function to invite the users to the channel - it will require a reintepretation of the userID to be transformed in a tag
-func inviteUsersToChannel(channelID string, users []string, commandUserID string) error {
-	slackToken := os.Getenv("SLACK_BOT_TOKEN")
-	if slackToken == "" {
-		return fmt.Errorf("Slack token not set")
-	}
-	inviteApiUrl := "https://slack.com/api/conversations.invite"
-
-	// Ensure the command user is always invited
-	users = append(users, commandUserID)
-
-	// ✅ Convert ONLY usernames to Slack user IDs
-	userIDs := convertUserTagsToIDs(users)
-
-	// ✅ Add the command user ID directly (since we already have it)
-	userIDs = append(userIDs, commandUserID)
-
-	// Remove duplicate user IDs
-	userIDSet := make(map[string]bool)
-	var uniqueUserIDs []string
-	for _, userID := range userIDs {
-		if _, exists := userIDSet[userID]; !exists {
-			userIDSet[userID] = true
-			uniqueUserIDs = append(uniqueUserIDs, userID)
-		}
-	}
-	if len(uniqueUserIDs) == 0 {
-		return fmt.Errorf("No valid users to invite")
-	}
-
-	//creating comma-separated users IDs for the payload
-	payload := map[string]interface{}{
-		"channel": channelID,
-		"users":   strings.Join(userIDs, ","),
-	}
-
-	payloadBytes, _ := json.Marshal(payload)
-
-	//Creating the HTTP request
-	req, err := http.NewRequest("POST", inviteApiUrl, bytes.NewBuffer(payloadBytes))
-	if err != nil {
-		return err
-	}
-
-	//headers setting
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+slackToken)
-
-	//Executing the request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	//decoding the response
-	var slackResp map[string]interface{}
-	json.NewDecoder(resp.Body).Decode(&slackResp)
-
-	//Printing Slack Response - just for Debugging
-	fmt.Println("DEBUG - Invite Users Response:", slackResp)
-
-	if ok, exists := slackResp["ok"].(bool); !exists || !ok {
-		return fmt.Errorf("Slack API error: %v", slackResp["error"])
-	}
-
-	log.Printf("✅ Invited users %v to channel ID: %s", userIDs, channelID)
-	return nil
-
-}
-
-// creating the function to convert the @usernames to user IDs
-func convertUserTagsToIDs(userTags []string) []string {
-	slackToken := os.Getenv("SLACK_BOT_TOKEN")
-	if slackToken == "" {
-		log.Println("Slack token not set")
-		return nil
-	}
-
-	url := "https://slack.com/api/users.list"
-
-	req, _ := http.NewRequest("GET", url, nil)
-	req.Header.Set("Authorization", "Bearer "+slackToken)
-	req.Header.Set("Content-Type", "application/json")
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Println("❌ Error fetching user list:", err)
-		return nil
-	}
-	defer resp.Body.Close()
-
-	var slackResp map[string]interface{}
-	json.NewDecoder(resp.Body).Decode(&slackResp)
-
-	if ok, exists := slackResp["ok"].(bool); !exists || !ok {
-		log.Println("⚠️ Slack API error:", slackResp["error"])
-		return nil
-	}
-
-	// Extract users from response
-	members, exists := slackResp["members"].([]interface{})
-	if !exists {
-		log.Println("❌ Error: No members found in workspace.")
-		return nil
-	}
-
-	// Build a map of username -> user ID
-	userMap := make(map[string]string)
-	for _, member := range members {
-		userData := member.(map[string]interface{})
-		username := userData["name"].(string)
-		userID := userData["id"].(string)
-
-		userMap[username] = userID
-	}
-
-	// Convert `@username` tags to Slack user IDs
-	var userIDs []string
-	for _, tag := range userTags {
-		username := strings.TrimPrefix(tag, "@")
-		if userID, exists := userMap[username]; exists {
-			userIDs = append(userIDs, userID)
-		} else {
-			log.Printf("⚠️ Warning: No Slack user found for @%s", username)
-		}
-	}
-
-	log.Println("DEBUG - Converted User IDs:", userIDs)
-	return userIDs
-}
-
-/*
-After a few tries, the bot still failed - the channel gets created correctly, but the bot fails the invitation
-
-	yorodinson@YorOdinSon ~ % curl -X POST "https://slack.com/api/conversations.invite" \
-	     -H "Authorization: Bearer xoxb- \
-	     -H "Content-Type: application/json" \
-	     -d '{"channel":"C08EE7BU5MH","users":"U12345,U67890"}'
-
-{"ok":false,"error":"not_in_channel","warning":"missing_charset","response_metadata":{"warnings":["missing_charset"]}}%
-
-It turns out, the bot doesn't join the channel so it can't invite people in.
-("error":"not_in_channel")
-
-Creating the function to let the Bot join the channel
-*/
-func joinChannel(channelID string) error {
-	slackToken := os.Getenv("SLACK_BOT_TOKEN")
-	if slackToken == "" {
-		return fmt.Errorf("Slack token not set")
-	}
-
-	url := "https://slack.com/api/conversations.join"
-
-	payload := map[string]string{
-		"channel": channelID,
-	}
-
-	payloadBytes, _ := json.Marshal(payload)
-
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(payloadBytes))
-	if err != nil {
-		return err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+slackToken)
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	var slackResp map[string]interface{}
-	json.NewDecoder(resp.Body).Decode(&slackResp)
-
-	// Debugging: Print response
-	fmt.Println("DEBUG - Join Channel Response:", slackResp)
-
-	if ok, exists := slackResp["ok"].(bool); !exists || !ok {
-		return fmt.Errorf("Slack API error: %v", slackResp["error"])
-	}
-
-	log.Printf("✅ Bot joined channel ID: %s", channelID)
-	return nil
-}
-
-/*
-parsing a little bit - I'll be expecting a string like
-
-/ninjaescal case: https://<link-to-the-case> client: <client-name> invite: @<user.tag1> @<user.tag2>...
-
-So I'll need to parse everything that comes after "case:" for the case URL from which I'll be creating the case number, then after "client: " for a <client-name>, and lastly after "invite:" for the invitee
-*/
 
 func parseCommand(input string) (string, string, []string) {
 	caseRegex := regexp.MustCompile(`case:\s*(https?://[^\s]+)`)
@@ -367,9 +78,14 @@ func handleSlackCommand(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintln(w, "⚠️ Missing required fields: `client:`")
 		return
 	}
-
+	//working out the case AND the jiraAPIOutcome
 	jiraCase := extractJira(caseURL)
-	channelName := fmt.Sprintf("tmp-%s-escalation-%s", clientName, jiraCase)
+
+	//API outcome ingesting
+
+	jiraIssue := getFromJira(caseURL)
+
+	channelName := fmt.Sprintf("escalation-%s-%s-temp", jiraCase, clientName)
 
 	// Fix invalid characters
 	channelName = fixChannelName(channelName)
@@ -386,6 +102,14 @@ func handleSlackCommand(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		fmt.Fprintf(w, "⚠️ Channel created, but bot failed to join: %s\n", err)
 		return
+	}
+
+	//✅ First I'll create the message, then I'll invite Users in
+	err = sendSlackMessage(channelID, jiraIssue, caseURL)
+	if err != nil {
+		fmt.Fprintf(w, "⚠️ Channel created, bot joined, but failed to send message: %s\n", err)
+	} else {
+		fmt.Fprintf(w, "✅ message sent!")
 	}
 
 	// ✅ Always invite the command user along with any specified users
